Revert 8779960..b61f79a
This rolls back to commit 8779960c5410f4422cff49ea078c7d1cff091220.

+++ Game/Sensors.cpp
@@ -25,7 +25,7 @@ int maxTurnRight = 0;
 int minTurnRight = 1023;
 
 void calibrate() {
-    unsigned long time = millis();
+    int time = millis();
     int leftSensor, rightSensor;
     int leftPhotoSensor, rightPhotoSensor;
     int leftTurnSensor, rightTurnSensor;
